GameBoardLines - TDD setup refactoring the TTT hash lines
