+++ src/components/game-board/GameBoard.test.js
@@ -49,46 +49,6 @@ describe('Calculated Lines', () => {
   });
 });
 
-describe('Actual SVG Lines', () => {
-  const wrapper = shallow(<GameBoard />);
-  const lines = wrapper.find('line');
-  it('should have correct first vertical line', () => {
-    expect(lines.at(0).props()).toEqual({
-      x1: (500 / 3),
-      y1: 0,
-      x2: (500 / 3),
-      y2: 500,
-    });
-  });
-
-  it('should have correct second vertical line', () => {
-    expect(lines.at(1).props()).toEqual({
-      x1: (500 / 3) * 2,
-      y1: 0,
-      x2: (500 / 3) * 2,
-      y2: 500,
-    });
-  });
-
-  it('should have correct first horizontal line', () => {
-    expect(lines.at(2).props()).toEqual({
-      x1: 0,
-      y1: 500 / 3,
-      x2: 500,
-      y2: 500 / 3,
-    });
-  });
-
-  it('should have correct second horizontal line', () => {
-    expect(lines.at(3).props()).toEqual({
-      x1: 0,
-      y1: (500 / 3) * 2,
-      x2: 500,
-      y2: (500 / 3) * 2,
-    });
-  });
-});
-
 it('should match a snapshot', () => {
   const wrapper = shallow(<GameBoard />);
   expect(wrapper).toMatchSnapshot();
